rename a file and add some comments
